requirements.txt
# Alpha Discovery API Dependencies

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Data Sources
requests>=2.28.0
yfinance>=0.2.0,<0.3.0
alpaca-trade-api>=3.0.0

# CORS
python-multipart>=0.0.6

# Environment variables (optional)
python-dotenv>=1.0.0
